chore(ast): split declarations into seperate file
